+++ libs/musicr/ui/admin-products/src/lib/components/product-form/product-form.component.ts
@@ -41,7 +41,7 @@ export function createProductForm(): FormGroup {
 })
 export class ProductFormComponent extends ApsFormComponent<Product> implements OnChanges {
   fileList: NzUploadFile[] = [];
-  allowedMimeType = ['image/jpeg', 'image/pjpeg', 'image/jpeg', 'image/pjpeg', 'image/png'];
+  allowedMimeType = ['image/jpeg', 'image/jpg'];
   previewImage: string | undefined = '';
   previewVisible = false;
   photosUrl: string;