+++ libs/valmira/ui/core/src/lib/entity.metadata.ts
@@ -2,7 +2,6 @@ import { EntityMetadataMap } from '@ngrx/data';
 
 const entityMetadata: EntityMetadataMap = {
   AdditionalProduct: {},
-  Category: {},
   Photo: {},
   Place: {},
   Promocode: {},