+++ libs/valmira/ui/place-search/src/lib/containers/place-search-container/place-search-container.component.ts
@@ -1,11 +1,9 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
-import { ApsListContainerComponent, filterNil } from '@arphase/ui/core';
+import { ApsListContainerComponent } from '@arphase/ui/core';
 import { QueryParams } from '@ngrx/data';
-import { select, Store } from '@ngrx/store';
+import { Store } from '@ngrx/store';
 import { Place } from '@valmira/domain';
-import { fromPlaces, PlaceCollectionService, PlaceDataService } from '@valmira/ui/places/data';
-import { keyBy } from 'lodash';
-import { map } from 'rxjs';
+import { PlaceCollectionService, PlaceDataService } from '@valmira/ui/places/data';
 
 @Component({
   selector: 'vma-place-search-container',
@@ -14,11 +12,6 @@ import { map } from 'rxjs';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class PlaceSearchContainerComponent extends ApsListContainerComponent<Place> implements OnInit {
-  summary$ = this.store.pipe(
-    select(fromPlaces.selectors.getPlacesCategorySummary),
-    filterNil(),
-    map(summary => keyBy(summary, 'category'))
-  );
   constructor(
     protected placeCollectionService: PlaceCollectionService,
     protected placeDataService: PlaceDataService,
@@ -30,7 +23,6 @@ export class PlaceSearchContainerComponent extends ApsListContainerComponent<Pla
   ngOnInit() {
     const params = { pageSize: String(50), onlyActives: String(true) };
     this.placeCollectionService.getWithQuery(params);
-    this.store.dispatch(fromPlaces.actions.getCategorySummary({ params }));
   }
 
   filterItems(queryParams?: QueryParams): void {
@@ -39,7 +31,6 @@ export class PlaceSearchContainerComponent extends ApsListContainerComponent<Pla
       ...queryParams,
       resetList: String(true),
     };
-    this.store.dispatch(fromPlaces.actions.getCategorySummary({ params }));
     this.placeCollectionService.getWithQuery(params);
   }
 }