+++ libs/valmira/ui/place-search/src/lib/place-search.module.ts
@@ -16,7 +16,6 @@ import { NzSkeletonModule } from 'ng-zorro-antd/skeleton';
 import { PlaceSearchFormComponent } from './components/place-search-form/place-search-form.component';
 import { PlaceSearchComponent } from './components/place-search/place-search.component';
 import { PlaceSearchContainerComponent } from './containers/place-search-container/place-search-container.component';
-import { SumAllPlacesPipe } from './pipes/sum-all-places.pipe';
 import { PlaceSearchRoutingModule } from './place-search-routing.module';
 
 @NgModule({
@@ -37,6 +36,6 @@ import { PlaceSearchRoutingModule } from './place-search-routing.module';
     PlaceSearchRoutingModule,
     ReactiveFormsModule,
   ],
-  declarations: [PlaceSearchContainerComponent, PlaceSearchComponent, PlaceSearchFormComponent, SumAllPlacesPipe],
+  declarations: [PlaceSearchContainerComponent, PlaceSearchComponent, PlaceSearchFormComponent],
 })
 export class PlaceSearchModule {}