+++ libs/valmira/ui/places/data/src/lib/services/place-data.service.ts
@@ -1,8 +1,8 @@
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { ApsDataService, buildQueryParams } from '@arphase/ui/core';
-import { HttpUrlGenerator, QueryParams } from '@ngrx/data';
-import { Place, PlaceCategorySummary } from '@valmira/domain';
+import { ApsDataService } from '@arphase/ui/core';
+import { HttpUrlGenerator } from '@ngrx/data';
+import { Place } from '@valmira/domain';
 import { Observable } from 'rxjs';
 
 @Injectable({ providedIn: 'root' })
@@ -16,10 +16,5 @@ export class PlaceDataService extends ApsDataService<Place> {
   getOccupiedDates(id: number, dateType: string): Observable<Date[]> {
     const params = new HttpParams({ fromObject: { dateType } });
     return this.http.get<Date[]>(`/vmaApi/places/${id}/occupied-dates`, { params });
-  }
-
-  getCategorySummary(queryParams: QueryParams): Observable<PlaceCategorySummary> {
-    const params = buildQueryParams(queryParams);
-    return this.http.get<PlaceCategorySummary>(`/vmaApi/places/count/category`, { params });
   }
 }