+++ libs/valmira/ui/places/data/src/lib/state/places.actions.ts
@@ -1,6 +1,4 @@
-import { QueryParams } from '@ngrx/data';
 import { createAction, props } from '@ngrx/store';
-import { PlaceCategorySummary } from '@valmira/domain';
 
 export const getStartOccupiedDates = createAction('[Places] Get start date occupied dates', props<{ id: number }>());
 
@@ -19,12 +17,3 @@ export const getEndDateOccupiedDatesSuccess = createAction(
 );
 
 export const getEndDateOccupiedDatesFailed = createAction('[Places] Get end date occupied dates failed');
-
-export const getCategorySummary = createAction('[Places] Get category summary', props<{ params: QueryParams }>());
-
-export const getCategorySummarySuccess = createAction(
-  '[Places] Get category summary success',
-  props<{ summary: PlaceCategorySummary }>()
-);
-
-export const getCategorySummaryFailed = createAction('[Places] Get category summary failed');